Idempotent networking setup, other minor fixes.
The host only networking interface name has a prefix appended to the key
that comes back from Machine.HostonlyNets & virtualbox.DHCPs. Using
these to match the current configured networking entities for
idempotency.

Hard-linking the image file because Virtualbox likes to delete it along
with the virtual machine. This is usually not what you want for a cached
cloud image.

Starting the machine after creating/matching it.

Conflicts:
	main.go