Small refactor.
I moved the disk clone to the top of the machine creation; and am runing
the clone in a seprate go routine, then waiting for it to complete
before attaching it to the machine. There really isn't a speedup; but at
least things can happen while the disk is being cloned.